Queue: +presentation; Draw: swapchain and frames-in-flight
drawFrame no longer needs to wait for the graphics queue submission to
complete to make the presentation call. Presentation is now
asynchronously handlend by the managed queue, and it correctly
distinguishes if the presentation queue is the same queue or not.

The whole swapchain image handling and frames-in-flight limiting has
been reworked to be simpler, more asynchronous, and avoid having reduced
frames-in-flight while resizing the window.

There is a bug that forces a reduced maximum of acquired swapchain
images. It causes the validation layer to complain while resizing the
window: "vkAcquireNextImageKHR: Application has already previously
acquired 2 images from swapchain. ..." Not sure if this is my bug or the
fault of MoltenVk or validation layer.
The workaround is in Vulkyrie.Engine.Main:117.